# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: QProv API
  contact:
    name: GitHub Repository
    url: https://github.com/UST-QuAntiL/QProv
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.3.0
servers:
- url: http://localhost:5020/qprov
  description: Generated server url
paths:
  /virtual-machines/{virtualMachineId}/characteristics:
    get:
      tags:
      - virtual-machine
      description: "Retrieve the calibration characteristics from the given gate.\
        \ By using the latest parameter only the latest data is retrieved, otherwise\
        \ all available data."
      operationId: getHardwareCharacteristics
      parameters:
      - name: virtualMachineId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: latest
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelHardwareCharacteristicsDto'
        "404":
          description: No characteristics for this Gate available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelHardwareCharacteristicsDto'
    put:
      tags:
      - virtual-machine
      description: "Create a new VirtualMachine and return the link which can then\
        \ be used to retrieve, update, and delete it."
      operationId: addHardwareCharacteristics
      parameters:
      - name: virtualMachineId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HardwareCharacteristicsDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelVirtualMachineDto'
...
  /virtual-machines:
    get:
      tags:
      - virtual-machine
      description: Retrieve all classical hardware virtual machines.
      operationId: getVirtualMachines
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelVirtualMachineDto'
    post:
      tags:
      - virtual-machine
      description: "Create a new VirtualMachine and return the link which can then\
        \ be used to retrieve, update, and delete it."
      operationId: createVirtualMachine
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VirtualMachineDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelVirtualMachineDto'
...
  /virtual-machines/{virtualMachineId}:
    get:
      tags:
      - virtual-machine
      description: Retrieve a specific VirtualMachine and its basic properties.
      operationId: getVirtualMachine
      parameters:
      - name: virtualMachineId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelVirtualMachineDto'
        "404":
          description: Not Found. VirtualMachine with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelVirtualMachineDto'
...
  /providers/{providerId}:
    get:
      tags:
      - provider
      description: Retrieve a specific provider and its basic properties.
      operationId: getProvider
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelProviderDto'
        "404":
          description: Not Found. Provider with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelProviderDto'
  /providers/{providerId}/qpus:
    get:
      tags:
      - provider
      description: Retrieve all QPUs of the provider.
      operationId: getQPUs
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelQpuDto'
        "404":
          description: Provider with the ID not available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelQpuDto'
  /providers/{providerId}/qpus/{qpuId}:
    get:
      tags:
      - provider
      description: Retrieve a specific QPU and its basic properties.
      operationId: getQPU
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelQpuDto'
        "400":
          description: QPU belongs not to specified provider.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelQpuDto'
        "404":
          description: Not Found. QPU with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelQpuDto'
  /providers/{providerId}/qpus/{qpuId}/qubits:
    get:
      tags:
      - provider
      description: Retrieve all Qubits of the QPU.
      operationId: getQubits
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelQubitDto'
        "404":
          description: Provider or QPU with the ID not available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelQubitDto'
...
  /providers/{providerId}/qpus/{qpuId}/aggregated-data:
    get:
      tags:
      - provider
      operationId: getLinksToAggregatedData
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/RepresentationModelObject'
        "404":
          description: Provider or QPU not found.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/RepresentationModelObject'
...
components:
  schemas:
    HardwareCharacteristicsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        recordingTime:
          type: string
          format: date-time
        cpuUsage:
          type: number
          format: float
        clockSpeed:
          type: number
          format: float
        ramUsage:
          type: number
          format: float
        diskUsage:
          type: number
          format: float
    EntityModelVirtualMachineDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        cpu:
          type: string
        cpuCores:
          type: integer
          format: int32
        ramSize:
          type: integer
          format: int32
        diskSize:
          type: integer
          format: int32
        _links:
          $ref: '#/components/schemas/Links'
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
    VirtualMachineDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        cpu:
          type: string
        cpuCores:
          type: integer
          format: int32
        ramSize:
          type: integer
          format: int32
        diskSize:
          type: integer
          format: int32
...
    CollectionModelEntityModelVirtualMachineDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            virtualMachineDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelVirtualMachineDto'
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelHardwareCharacteristicsDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            hardwareCharacteristicsDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelHardwareCharacteristicsDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelHardwareCharacteristicsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        recordingTime:
          type: string
          format: date-time
        cpuUsage:
          type: number
          format: float
        clockSpeed:
          type: number
          format: float
        ramUsage:
          type: number
          format: float
        diskUsage:
          type: number
          format: float
        _links:
          $ref: '#/components/schemas/Links'
...
    EntityModelProviderDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        offeringURL:
          type: string
          format: url
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelQpuDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            qpuDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelQpuDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelQpuDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        lastUpdated:
          type: string
          format: date-time
        lastCalibrated:
          type: string
          format: date-time
        maxShots:
          type: integer
          format: int32
        queueSize:
          type: integer
          format: int32
        numberOfQubits:
          type: integer
          format: int32
        avgT1Time:
          type: number
        avgT2Time:
          type: number
        avgReadoutError:
          type: number
        avgMultiQubitGateError:
          type: number
        avgSingleQubitGateError:
          type: number
        avgMultiQubitGateTime:
          type: number
        avgSingleQubitGateTime:
          type: number
        maxGateTime:
          type: number
        simulator:
          type: boolean
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelQubitDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            qubitDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelQubitDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelQubitDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        _links:
          $ref: '#/components/schemas/Links'
...
    RepresentationModelObject:
      type: object
      properties:
        _links:
          $ref: '#/components/schemas/Links'
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean